# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/send-otp": {
      "post": {
        "operationId": "AuthController_sendOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP accepted for delivery",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "OTP sent successfully",
                    "expiresIn": 300,
                    "resendAfter": 30
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid phone number"
          },
          "429": {
            "description": "Rate limit, cooldown, or lock active"
          }
        },
        "summary": "Send OTP to an Indian mobile number",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified. Existing users receive tokens; new users continue to registration.",
            "content": {
              "application/json": {
                "schema": {
                  "examples": {
                    "existingUser": {
                      "value": {
                        "needsRegistration": false,
                        "user": {
                          "id": "0bb3c81a-cb04-42a9-9414-7f362a5bb143",
                          "phone": "9876543210",
                          "email": "guest@quicknest.in",
                          "globalRole": "USER",
                          "isPhoneVerified": true,
                          "isEmailVerified": false,
                          "status": "active",
                          "createdAt": "2026-04-26T00:00:00.000Z",
                          "updatedAt": "2026-04-26T00:00:00.000Z"
                        },
                        "tokens": {
                          "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                          "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                          "tokenType": "Bearer",
                          "expiresIn": 900
                        }
                      }
                    },
                    "newUser": {
                      "value": {
                        "needsRegistration": true,
                        "message": "Phone verified successfully",
                        "verificationToken": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Expired or incorrect OTP"
          },
          "429": {
            "description": "OTP verification locked"
          }
        },
        "summary": "Verify OTP and login existing users or continue registration",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered and tokens issued"
          },
          "400": {
            "description": "Invalid verification token or payload is invalid"
          },
          "409": {
            "description": "Phone or email already exists"
          }
        },
        "summary": "Register a user after verification",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credentials accepted and tokens issued"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "429": {
            "description": "Too many login attempts"
          }
        },
        "summary": "Login with email and password fallback",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/firebase": {
      "post": {
        "operationId": "AuthController_firebaseLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirebaseLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Firebase verified user logged in or registration required"
          },
          "400": {
            "description": "Invalid Firebase token"
          }
        },
        "summary": "Verify Firebase phone auth token and login or continue registration",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh-token": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refresh token rotated successfully"
          },
          "401": {
            "description": "Invalid, revoked, or expired refresh token"
          },
          "429": {
            "description": "Too many refresh attempts"
          }
        },
        "summary": "Rotate refresh token and issue a new access token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refresh token revoked",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Logged out successfully"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token"
          }
        },
        "summary": "Revoke the current refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "QuickNest API",
    "description": "Backend APIs for QuickNest hotel booking platform",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SendOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "9876543210",
            "description": "10 digit Indian mobile number"
          }
        },
        "required": [
          "phone"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "9876543210",
            "description": "10 digit Indian mobile number"
          },
          "otp": {
            "type": "string",
            "example": "123456",
            "description": "6 digit OTP sent to the mobile number"
          }
        },
        "required": [
          "phone",
          "otp"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "verificationToken": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
            "description": "Short-lived verification token from OTP or Firebase verification"
          },
          "email": {
            "type": "string",
            "example": "guest@quicknest.in",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "example": "QuickNest@123",
            "minLength": 8,
            "maxLength": 128,
            "description": "Must include uppercase, lowercase, number, and special character"
          }
        },
        "required": [
          "verificationToken",
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "guest@quicknest.in",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "example": "QuickNest@123",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "FirebaseLoginDto": {
        "type": "object",
        "properties": {}
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token issued by login, register, or refresh-token",
            "minLength": 20
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Current refresh token to revoke",
            "minLength": 20
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
    }
  }
}